+++ CHANGELOG.md
@@ -32,6 +32,10 @@
 - BREAKING FEATURE: update `section_fileinfo()` to show non-empty filetype even in not normal buffers (like plugin's scratch buffers, help, quickfix, etc.). Previously it showed nothing, which was a mistake as filetype can be a valuable information.
 - FEATURE: prefer using 'mini.icons' as icon provider for `section_fileinfo()`.
 
+## mini.tabline
+
+- FEATURE: prefer using 'mini.icons' as icon provider.
+
 
 # Version 0.13.0